+++ Vonk.Plugin.DocumentOperation.Test/DocumentOperationTests.cs
@@ -15,8 +15,9 @@ using System.Collections.Generic;
 using Vonk.Fhir.R3;
 using Vonk.Core.Context.Features;
 using Vonk.Core.Support;
-using static Hl7.Fhir.Model.Bundle;
 using System.Linq;
+using Vonk.Core.ElementModel;
+using Hl7.Fhir.Specification;
 
 namespace Vonk.Plugin.DocumentOperation.Test
 {
@@ -26,7 +27,7 @@ namespace Vonk.Plugin.DocumentOperation.Test
             - (X) $document should persist the generated document on request
             - (X) $document should return an INVALID_REQUEST when being called with POST and an missing id
             - (X) $document should throw an internal server error if a local reference to a resource, which should be included in the document, can't be found.
-            - (X) $document should thrw an internal server error if an external reference is requested to be included in the document
+            - (X) $document should throw an internal server error if an external reference is requested to be included in the document
     */
 
     public class DocumentOperationTests
@@ -36,10 +37,11 @@ namespace Vonk.Plugin.DocumentOperation.Test
         private ILogger<DocumentService> _logger = Logger<DocumentService>();
         private Mock<ISearchRepository> _searchMock = new Mock<ISearchRepository>();
         private Mock<IResourceChangeRepository> _changeMock = new Mock<IResourceChangeRepository>();
+        private IStructureDefinitionSummaryProvider _schemaProvider = new PocoStructureDefinitionSummaryProvider();
 
         public DocumentOperationTests()
         {
-            _documentService = new DocumentService(_searchMock.Object, _changeMock.Object, _logger);
+            _documentService = new DocumentService(_searchMock.Object, _changeMock.Object, _schemaProvider, _logger);
         }
 
         [Fact]
@@ -66,8 +68,8 @@ namespace Vonk.Plugin.DocumentOperation.Test
             // Check response status
             testContext.Response.HttpResult.Should().Be(StatusCodes.Status200OK, "$document should succeed with HTTP 200 - OK on test composition");
             testContext.Response.Payload.Should().NotBeNull();
-            var document = testContext.Response.Payload.ToPoco<Bundle>();
-            document.Type.Should().Be(BundleType.Document);
+            var bundleType = testContext.Response.Payload.SelectText("type");
+            bundleType.Should().Be("document");
         }
 
         [Fact]
@@ -101,8 +103,8 @@ namespace Vonk.Plugin.DocumentOperation.Test
             // Check response status
             testContext.Response.HttpResult.Should().Be(StatusCodes.Status200OK, "$document should succeed with HTTP 200 - OK on test composition");
             testContext.Response.Payload.Should().NotBeNull();
-            var document = testContext.Response.Payload.ToPoco<Bundle>();
-            document.Type.Should().Be(BundleType.Document);
+            var bundleType = testContext.Response.Payload.SelectText("type");
+            bundleType.Should().Be("document");
         }
 
         [Fact]
@@ -209,6 +211,7 @@ namespace Vonk.Plugin.DocumentOperation.Test
 
             // Check response status
             testContext.Response.HttpResult.Should().Be(StatusCodes.Status500InternalServerError, "$document should return HTTP 500 - Internal Server error when a reference which is referenced by the composition can't be resolved");
+            testContext.Response.Outcome.Issue.Count(issue => issue.Code == OperationOutcome.IssueType.NotFound).Should().NotBe(0, "OperationOutcome should explicitly mention that the reference could not be found");
         }
 
         [Fact]
@@ -242,6 +245,7 @@ namespace Vonk.Plugin.DocumentOperation.Test
 
             // Check response status
             testContext.Response.HttpResult.Should().Be(StatusCodes.Status500InternalServerError, "$document should return HTTP 500 - Internal Server error when a reference which is referenced by the composition can't be resolved");
+            testContext.Response.Outcome.Issue.Count(issue => issue.Code == OperationOutcome.IssueType.NotFound).Should().NotBe(0, "OperationOutcome should explicitly mention that the reference could not be found");
         }
 
         [Fact]
@@ -279,13 +283,14 @@ namespace Vonk.Plugin.DocumentOperation.Test
 
             // Check response status
             testContext.Response.HttpResult.Should().Be(StatusCodes.Status500InternalServerError, "$document should return HTTP 500 - Internal Server error when a reference which is referenced by the composition can't be resolved");
+            testContext.Response.Outcome.Issue.Count(issue => issue.Code == OperationOutcome.IssueType.NotFound).Should().NotBe(0, "OperationOutcome should explicitly mention that the reference could not be found");
         }
 
         [Fact]
         public async Task DocumentOperationSuccessCompleteComposition()
         {
             // Setup Composition resource
-            var composition = CreateTestCompositionInclList(); // Unresolvable reference (Medication resource) in MedicationStatement resource (4. level)
+            var composition = CreateTestCompositionInclList();
             var compositionSearchResult = new SearchResult(new List<IResource>() { composition }, 1, 1);
 
             var list = CreateTestList();
@@ -317,6 +322,7 @@ namespace Vonk.Plugin.DocumentOperation.Test
 
             // Check response status
             testContext.Response.HttpResult.Should().Be(StatusCodes.Status200OK, "$document should return HTTP 200 - OK when all references in the composition (incl. recursive references) can be resolved");
+            testContext.Response.Payload.SelectNodes("entry.resource").Count().Should().Be(4, "Expected Composition, List, MedicationStatement and Medication resources to be in the document");
         }
 
         [Fact]
@@ -342,6 +348,41 @@ namespace Vonk.Plugin.DocumentOperation.Test
 
             // Check response status
             testContext.Response.HttpResult.Should().Be(StatusCodes.Status500InternalServerError, "$document should return HTTP 500 - Internal Server error when an external reference is referenced by the composition");
+            testContext.Response.Outcome.Issue.Count(issue => issue.Code == OperationOutcome.IssueType.NotSupported).Should().NotBe(0, "OperationOutcome should highlight that this feature is not supported");
+        }
+
+        [Fact]
+        public async Task DocumentBundleContainsIdentifier()
+        {
+            // Setup Composition resource
+            var composition = CreateTestCompositionNoReferences();
+            var searchResult = new SearchResult(new List<IResource>() { composition }, 1, 1);
+            _searchMock.Setup(repo => repo.Search(It.IsAny<IArgumentCollection>(), It.IsAny<SearchOptions>())).ReturnsAsync(searchResult);
+
+            // Create VonkContext for $document (GET / Instance level)
+            var testContext = new VonkTestContext(VonkInteraction.instance_custom);
+            testContext.Arguments.AddArguments(new[]
+            {
+                new Argument(ArgumentSource.Path, ArgumentNames.resourceType, "Composition"),
+                new Argument(ArgumentSource.Path, ArgumentNames.resourceId, "test")
+            });
+            testContext.TestRequest.CustomOperation = "document";
+            testContext.TestRequest.Method = "GET";
+
+            // Execute $document
+            await _documentService.DocumentInstanceGET(testContext);
+
+            testContext.Response.HttpResult.Should().Be(StatusCodes.Status200OK, "$document should succeed with HTTP 200 - OK on test composition");
+            testContext.Response.Payload.Should().NotBeNull();
+
+            var identifier = testContext.Response.Payload.SelectNodes("identifier");
+            identifier.Should().NotBeEmpty("A document SHALL contain at least one identifier");
+        }
+
+        [Fact]
+        public async Task DocumentOperationCanIncludeCustomResources()
+        {
+
         }
 
         // $document is expected to fail if a resource reference is missing, this should be checked on all levels of recursion.
@@ -370,6 +411,11 @@ namespace Vonk.Plugin.DocumentOperation.Test
             composition.Section.Add(sectionComponent);
 
             return composition.ToIResource();
+        }
+
+        private IResource CreateTestCompositionInclCustomResource()
+        {
+            return new Composition() { Id = "test", VersionId = "v1", Subject = new ResourceReference("CustomResourceTest/test") }.ToIResource();
         }
 
         private IResource CreateTestPatient()